requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
python-dotenv==1.0.1
dj-rest-auth==5.0.2
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0